static ==> notvar
Our preinterpretations should not use both static and notvar
so confusingly (even if notvar is only used as an annotation
in constructor modes, e.g. [notvar|any]).
Anyway, the only true meaning of what we used to call static
is notvar, that's what we care for in the analysis for Narradar